+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.9.5
 
  - Fix castp permission (target player doesn't need permission, only command executor needs it)
+ - Add Factions support for build perms. (Kind of hacky at present- Factions, why u no API??)
 
 ## 2.9.4